Refactor C code into multiple files
