+++ src/util.h
@@ -1,0 +1,26 @@
+#ifndef UTIL_H
+#define UTIL_H
+
+//#define DEBUG
+#ifdef DEBUG
+#define debug(...) do { fprintf(stderr, __VA_ARGS__); fprintf(stderr, "\r\n"); } while(0)
+#else
+#define debug(...)
+#endif
+
+// MAC address utilities
+#define MACADDR_STR_LEN      18 // aa:bb:cc:dd:ee:ff and a null terminator
+
+int string_to_macaddr(const char *str, unsigned char *mac);
+int macaddr_to_string(const unsigned char *mac, char *str);
+
+// Encoding
+struct netif;
+void encode_kv_long(struct netif *nb, const char *key, long value);
+void encode_kv_ulong(struct netif *nb, const char *key, unsigned long value);
+void encode_kv_bool(struct netif *nb, const char *key, int value);
+void encode_string(char *buf, int *index, const char *str);
+void encode_kv_string(struct netif *nb, const char *key, const char *str);
+void encode_kv_macaddr(struct netif *nb, const char *key, const unsigned char *macaddr);
+
+#endif // UTIL_H
